requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
pygame>=2.1.0,<3.0.0
neat-python>=0.92
graphviz>=0.20.0
matplotlib>=3.5.0

# Optional development dependencies
# Install with: pip install -e ".[dev]"
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
# isort>=5.12.0
